ai-services/chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.3
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
